msm: adsprpc: Do not search the global map in the process-specific list
If a user makes the ioctl call for the fastrpc_internal_mmap with the
global map flag, fd, and va corresponding to some map already present
in the process-specific list, then this map present in the process-
specific list could be added to the global list. Because global maps
are also searched in the process-specific list. If a map gets removed
from the global list and another concurrent thread is using the same
map for a process-specific use case, it could lead to a use-after-free.
Avoid searching the global map in the process-specific list.

Change-Id: Ife06a3018332017f0076171fb4be98f3b0ed3502
Acked-by: Abhishek Singh <abhishes@qti.qualcomm.com>
Signed-off-by: Momojit Ghosh <momoghos@qti.qualcomm.com>
(cherry picked from commit 1d6d6e3aa9e0d59152c8b339839cb8f53e0a60a3)